test: ensure integration tests run single-threaded
