+++ tlspuffin/tests/mutations.rs
@@ -12,16 +12,15 @@ use puffin::libafl::state::StdState;
 use puffin::libafl_bolts::rands::{RomuDuoJrRand, StdRand};
 use puffin::test_utils::AssertExecution;
 use puffin::trace::{Action, Step, Trace};
-
-use crate::put_registry::tls_registry;
-use crate::query::TlsQueryMatcher;
-use crate::test_utils::default_runner_for;
-use crate::tls::fn_impl::{
+use tlspuffin::put_registry::tls_registry;
+use tlspuffin::query::TlsQueryMatcher;
+use tlspuffin::test_utils::default_runner_for;
+use tlspuffin::tls::fn_impl::{
     fn_client_hello, fn_encrypt12, fn_seq_1, fn_sign_transcript, fn_signature_algorithm_extension,
     fn_support_group_extension,
 };
-use crate::tls::seeds::_seed_client_attacker12;
-use crate::tls::TLS_SIGNATURE;
+use tlspuffin::tls::seeds::_seed_client_attacker12;
+use tlspuffin::tls::TLS_SIGNATURE;
 
 fn create_state() -> StdState<
     Trace<TlsQueryMatcher>,
